# app/Http/Controllers/Api/v1/UserController.php
<?php

namespace App\Http\Controllers\Api\v1;

use Illuminate\Http\Request;
use App\Http\Controllers\Controller;
use Illuminate\Support\Facades\Auth;

use App\User;
use function PHPSTORM_META\type;

class UserController extends Controller
{
    /**
     * Display a listing of the resource.
     *
     * @return \Illuminate\Http\Response
     *
     * @SWG\GET(
     *     path="/api/v1/users",
     *     @SWG\Response(response="200", description="Get all users"),
     *     security={ {"passport": {} } }
     * )
     */
    public function index()
    {
        $user = new User;
        return $user->with('accommodations')->get();
    }

    /**
     * Store a newly created resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return \Illuminate\Http\Response
     *
     * @SWG\POST(
     *     path="/api/v1/users",
     *     @SWG\Response(response="200", description="Create one user"),
     *     security={ {"passport": {} } }
     * )
     */
    public function store(Request $request)
    {
        $user = new User;
        return $user->create($request->all());
    }

    /**
     * Display the specified resource.
     *
     * @param  int  $userId
     * @return \Illuminate\Http\Response
     *
     * @SWG\GET(
     *     path="/api/v1/users/{id}",
     *     @SWG\Response(response="200", description="Get one user by id"),
     *     security={ {"passport": {} } }
     * )
     */
    public function show($userId)
    {
        $user = new User;
        if (Auth::user()->id == $userId || Auth::user()->roles == 1){
            return  $user->findOrFail($userId);
        } else {
            return $user->select("username")->where('id', $userId)->get();
        }
        return response()->json(null, 404);
    }

    /**
     * Update the specified resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @param  int  $userId
     * @return \Illuminate\Http\Response
     *
     * @SWG\PUT(
     *     path="/api/v1/users/{id}",
     *     @SWG\Response(response="200", description="Update one user by id"),
     *     security={ {"passport": {} } }
     * )
     */
    public function update(Request $request, $userId)
    {
        $user = new User;
        $user = $user->findOrFail($userId);
        $user->update($request->all());

        return $user;
    }

    /**
     * Remove the specified resource from storage.
     *
     * @param  int  $userId
     * @return \Illuminate\Http\Response
     *
     * @SWG\DELETE(
     *     path="/api/v1/users/{id}",
     *     @SWG\Response(response="200", description="Delete one user by id"),
     *     security={ {"passport": {} } }
     * )
     */
    public function destroy($userId)
    {
        $user = new User;
        $user->findOrFail($userId)->delete();

        return response()->json(null, 204);
    }
}