# storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"Accommodation CRUD","version":"1.0"},"paths":{"\/api\/v1\/accommodations":{"get":{"summary":"Display a listing of the resource.","responses":{"200":{"description":"Get all accommodations"}},"security":[{"passport":[]}]},"post":{"summary":"Store a newly created resource in storage.","responses":{"201":{"description":"Create one accommodation"}},"security":[{"passport":[]}]}},"\/api\/v1\/accommodations\/{id}":{"get":{"summary":"Display the specified resource.","responses":{"200":{"description":"Get one accommodation by id"}},"security":[{"passport":[]}]},"put":{"summary":"Update the specified resource in storage.","responses":{"200":{"description":"Update one accommodation by id"}},"security":[{"passport":[]}]},"delete":{"summary":"Remove the specified resource from storage.","responses":{"200":{"description":"Delete one accommodation by id"}},"security":[{"passport":[]}]}},"\/api\/v1\/candidates":{"get":{"summary":"Display a listing of the resource.","responses":{"200":{"description":"Get all candidates"}},"security":[{"passport":[]}]},"post":{"summary":"Store a newly created resource in storage.","responses":{"200":{"description":"Create one candidate"}},"security":[{"passport":[]}]}},"\/api\/v1\/candidates\/{id}":{"get":{"summary":"Display the specified resource.","responses":{"200":{"description":"Get one candidate by id"}},"security":[{"passport":[]}]},"put":{"summary":"Update the specified resource in storage.","responses":{"200":{"description":"Update one candidate by id"}},"security":[{"passport":[]}]},"delete":{"summary":"Remove the specified resource from storage.","responses":{"200":{"description":"Delete one candidate by id"}},"security":[{"passport":[]}]}},"\/api\/v1\/messages":{"get":{"summary":"Display a listing of the resource.","responses":{"200":{"description":"Get all messages"}},"security":[{"passport":[]}]},"post":{"summary":"Store a newly created resource in storage.","responses":{"200":{"description":"Create one message"}},"security":[{"passport":[]}]}},"\/api\/v1\/messages\/{id}":{"get":{"summary":"Display the specified resource.","responses":{"200":{"description":"Get one message by id"}},"security":[{"passport":[]}]},"put":{"summary":"Update the specified resource in storage.","responses":{"200":{"description":"Update one message by id"}},"security":[{"passport":[]}]},"delete":{"summary":"Remove the specified resource from storage.","responses":{"200":{"description":"Delete one message by id"}},"security":[{"passport":[]}]}},"\/api\/v1\/missions":{"get":{"summary":"Display a listing of the resource.","responses":{"200":{"description":"Get all missions"}},"security":[{"passport":[]}]},"post":{"summary":"Store a newly created resource in storage.","responses":{"200":{"description":"Create one mission"}},"security":[{"passport":[]}]}},"\/api\/v1\/missions\/{id}":{"get":{"summary":"Display the specified resource.","responses":{"200":{"description":"Get one mission by id"}},"security":[{"passport":[]}]},"put":{"summary":"Update the specified resource in storage.","responses":{"200":{"description":"Update one mission by id"}},"security":[{"passport":[]}]},"delete":{"summary":"Remove the specified resource from storage.","responses":{"200":{"description":"Delete one mission by id"}},"security":[{"passport":[]}]}},"\/api\/v1\/pictures":{"get":{"summary":"Display a listing of the resource.","responses":{"200":{"description":"Get all pictures"}},"security":[{"passport":[]}]},"post":{"summary":"Store a newly created resource in storage.","responses":{"200":{"description":"Create one picture"}},"security":[{"passport":[]}]}},"\/api\/v1\/pictures\/{id}":{"get":{"summary":"Display the specified resource.","responses":{"200":{"description":"Get one picture by id"}},"security":[{"passport":[]}]},"put":{"summary":"Update the specified resource in storage.","responses":{"200":{"description":"Update one picture by id"}},"security":[{"passport":[]}]},"delete":{"summary":"Remove the specified resource from storage.","responses":{"200":{"description":"Delete one picture by id"}},"security":[{"passport":[]}]}},"\/api\/v1\/users":{"get":{"summary":"Display a listing of the resource.","responses":{"200":{"description":"Get all users"}},"security":[{"passport":[]}]},"post":{"summary":"Store a newly created resource in storage.","responses":{"200":{"description":"Create one user"}},"security":[{"passport":[]}]}},"\/api\/v1\/users\/{id}":{"get":{"summary":"Display the specified resource.","responses":{"200":{"description":"Get one user by id"}},"security":[{"passport":[]}]},"put":{"summary":"Update the specified resource in storage.","responses":{"200":{"description":"Update one user by id"}},"security":[{"passport":[]}]},"delete":{"summary":"Remove the specified resource from storage.","responses":{"200":{"description":"Delete one user by id"}},"security":[{"passport":[]}]}}},"definitions":{},"components":{"securitySchemes":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"https","flows":{"password":{"authorizationUrl":"http:\/\/localhost\/oauth\/authorize","tokenUrl":"http:\/\/localhost\/oauth\/token","refreshUrl":"http:\/\/localhost\/token\/refresh","scopes":[]}}}}}}